Add decode_mapping helper and refactor
